+++ __fixtures__/misc/output-proto-amino/amino-interface-proto-optionality/misc/all_fields.ts
@@ -1099,8 +1099,8 @@ export const EncodingTestForDontOmit = {
     message.dOPeriods = object.dOPeriods?.map(e => Duration.fromPartial(e)) || [];
     message.protos = object.protos?.map(e => AccessConfig.fromPartial(e)) || [];
     message.dOProtos = object.dOProtos?.map(e => AccessConfig.fromPartial(e)) || [];
-    message.auths = object.auths?.map(e => (Any.fromPartial(e) as any)) || [];
-    message.dOAuths = object.dOAuths?.map(e => (Any.fromPartial(e) as any)) || [];
+    message.auths = object.auths?.map(e => (GlobalDecoderRegistry.fromPartial(e) as any)) || [];
+    message.dOAuths = object.dOAuths?.map(e => (GlobalDecoderRegistry.fromPartial(e) as any)) || [];
     message.dec = object.dec ?? "";
     message.dODec = object.dODec ?? "";
     message.decs = object.decs?.map(e => e) || [];
@@ -2270,8 +2270,8 @@ export const EncodingTestForOmit = {
     message.oPeriods = object.oPeriods?.map(e => Duration.fromPartial(e)) || [];
     message.protos = object.protos?.map(e => AccessConfig.fromPartial(e)) || [];
     message.oProtos = object.oProtos?.map(e => AccessConfig.fromPartial(e)) || [];
-    message.auths = object.auths?.map(e => (Any.fromPartial(e) as any)) || [];
-    message.oAuths = object.oAuths?.map(e => (Any.fromPartial(e) as any)) || [];
+    message.auths = object.auths?.map(e => (GlobalDecoderRegistry.fromPartial(e) as any)) || [];
+    message.oAuths = object.oAuths?.map(e => (GlobalDecoderRegistry.fromPartial(e) as any)) || [];
     message.dec = object.dec ?? "";
     message.oDec = object.oDec ?? "";
     message.decs = object.decs?.map(e => e) || [];