+++ __fixtures__/misc/output-proto-amino/amino-interface/types.ts
@@ -36,8 +36,8 @@ export interface TelescopeGeneratedCodec<
   encode: (message: T, writer?: IBinaryWriter | any) => IBinaryWriter | any;
   decode: (input: IBinaryReader | Uint8Array | any, length?: number) => T;
   fromPartial: (object: any) => T | any;
-  fromJSON: (object: any) => T | any;
-  toJSON: (message: T | any) => any;
+  fromJSON?: (object: any) => T | any;
+  toJSON?: (message: T | any) => any;
   fromSDK?: (sdk: SDK) => T;
   fromSDKJSON?: (object: any) => SDK;
   toSDK?: (message: T) => SDK;
@@ -93,6 +93,7 @@ export interface MsgData {
 export interface Attribute {
   key: string;
   value: string;
+  index: boolean;
 }
 export interface Event {
   type: string;