+++ __fixtures__/v-next/outputv2/osmosis/gamm/v1beta1/genesis.ts
@@ -190,7 +190,7 @@ export const GenesisState = {
       const tag = reader.uint32();
       switch (tag >>> 3) {
         case 1:
-          message.pools.push((Any(reader) as Any));
+          message.pools.push((Any.decode(reader, reader.uint32()) as Any));
           break;
         case 2:
           message.nextPoolNumber = reader.uint64();