+++ packages/parser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.4](https://github.com/cosmology-tech/telescope/compare/@cosmology/proto-parser@1.4.3...@cosmology/proto-parser@1.4.4) (2024-01-19)
+
+**Note:** Version bump only for package @cosmology/proto-parser
+
+
+
+
+
+## [1.4.3](https://github.com/cosmology-tech/telescope/compare/@cosmology/proto-parser@1.4.2...@cosmology/proto-parser@1.4.3) (2024-01-16)
+
+**Note:** Version bump only for package @cosmology/proto-parser
+
+## [1.4.2](https://github.com/cosmology-tech/telescope/compare/@cosmology/proto-parser@1.4.1...@cosmology/proto-parser@1.4.2) (2024-01-16)
+
+**Note:** Version bump only for package @cosmology/proto-parser
+
 ## [1.4.1](https://github.com/cosmology-tech/telescope/compare/@cosmology/proto-parser@1.4.0...@cosmology/proto-parser@1.4.1) (2024-01-08)
 
 **Note:** Version bump only for package @cosmology/proto-parser