+++ packages/telescope/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.12](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.11...@cosmology/telescope@1.4.12) (2024-01-23)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.11](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.10...@cosmology/telescope@1.4.11) (2024-01-23)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.10](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.9...@cosmology/telescope@1.4.10) (2024-01-23)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.9](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.8...@cosmology/telescope@1.4.9) (2024-01-19)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.8](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.7...@cosmology/telescope@1.4.8) (2024-01-19)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.7](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.6...@cosmology/telescope@1.4.7) (2024-01-19)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.6](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.5...@cosmology/telescope@1.4.6) (2024-01-16)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.5](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.4...@cosmology/telescope@1.4.5) (2024-01-16)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.4](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.3...@cosmology/telescope@1.4.4) (2024-01-16)
+
+**Note:** Version bump only for package @cosmology/telescope
+
 ## [1.4.3](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.2...@cosmology/telescope@1.4.3) (2024-01-08)
 
 **Note:** Version bump only for package @cosmology/telescope